chore(profile): Moved components
